+++ src/Ds/Bundle/PlatformBundle/DependencyInjection/Configuration.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Ds\Bundle\PlatformBundle\DependencyInjection;
+
+use Symfony\Component\Config\Definition\Builder\TreeBuilder;
+use Symfony\Component\Config\Definition\ConfigurationInterface;
+
+/**
+ * Class Configuration
+ */
+class Configuration implements ConfigurationInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getConfigTreeBuilder()
+    {
+        $treeBuilder = new TreeBuilder;
+        $rootNode = $treeBuilder->root('ds_platform');
+
+        return $treeBuilder;
+    }
+}