+++ src/Ds/Bundle/PlatformBundle/DsPlatformBundle.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Ds\Bundle\PlatformBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Class DsPlatformBundle
+ */
+class DsPlatformBundle extends Bundle
+{
+}